Move ApiMock helpers to seperate file and don't expose it
