+++ plugins/git-release-manager/src/features/PromoteRc/hooks/usePromoteRc.test.ts
@@ -18,12 +18,12 @@ import { renderHook, act } from '@testing-library/react-hooks';
 import { waitFor } from '@testing-library/react';
 
 import {
-  mockApiClient,
   mockCalverProject,
   mockReleaseCandidateCalver,
   mockUser,
 } from '../../../test-helpers/test-helpers';
 import { usePromoteRc } from './usePromoteRc';
+import { mockApiClient } from '../../../test-helpers/mock-api-client';
 
 jest.mock('@backstage/core-plugin-api', () => ({
   ...jest.requireActual('@backstage/core-plugin-api'),
