+++ plugins/git-release-manager/src/features/RepoDetailsForm/Owner.test.tsx
@@ -18,7 +18,6 @@ import React from 'react';
 import { render, waitFor, screen } from '@testing-library/react';
 
 import {
-  mockApiClient,
   mockCalverProject,
   mockSearchCalver,
   mockUser,
@@ -26,6 +25,7 @@ import {
 import { TEST_IDS } from '../../test-helpers/test-ids';
 import { useProjectContext } from '../../contexts/ProjectContext';
 import { Owner } from './Owner';
+import { mockApiClient } from '../../test-helpers/mock-api-client';
 
 jest.mock('react-router', () => ({
   useNavigate: jest.fn(),
